Added a delta parameter for assertEquals.
As explained in the previous commit, the delta gives the
compared Numbers a range that they can differ by and still
have the assertion pass. The tricky part to this added parameter
is sorting out if calling the function with three parameters is
supposed to use the first argument as the error string, or the
last as the delta. This is sorted out by seeing if converting
all the arguments to Number leads to any of them being NaN. If so,
then the first argument is used as the error string.